api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
PyJWT[crypto]==2.10.1
bcrypt==4.2.1
gunicorn==23.0.0
boto3>=1.35.0,<2.0.0
anthropic>=0.69.0,<1.0.0
openpyxl==3.1.5
requests==2.32.3
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
cryptography==44.0.1
croniter==6.0.0
resend==2.22.0
beautifulsoup4==4.13.3
PyYAML==6.0.2
Jinja2==3.1.6
langgraph==0.6.11
langchain-anthropic==0.3.22
langchain-core==0.3.83
langsmith==0.4.37
